Fixed bug where favorite state wasn't being updated on re-render
This bug caused previously unfavorited movies to be favorited automatically. This is because favorite state was only being initialized on initial Movie component render, but ther was no logic to update the state whenever movie_ prop changed but the component Movie stayed the same. I solved it by adding a useEffect that updates the favorite state whenever the movie_ props changes