Updated stylings for genre in Genres.css
To get the behavior I wanted, I created an absolutely positioned child element called genreInner which is set to visbility hidden inside each genre, and made each genre relatively positioned. This makes it so that surrounding divs don't shift when hovering over a genre. On genre hover, I make the genre hidden and the genreInner visible.