Changed the stylings for the modal .contentWrap and .movieModalWrap
I changed the .contentWRap to position: absolute and added a transform line to center the content. The content was centered before this change, but the width was taking up 100% of the screen, and thus any click on the extra width would not close the modal. With this change, the modal now closes on clicking outside of the content.