Improved the modal display
I switched the overview and genres so that the genres is at the bottom and overview at the top. This makes more sense since all the other movie information is at the bottom as well. I also moved the overflow: auto to be in the overview paragraph rather than the overview to account for this change. Finally, I added a margin-top to the genres and padding to the individual genres